services/transformer/requirements.txt
# services/transformer/requirements.txt
mysql-connector-python==8.0.26
kafka-python==2.0.2
boto3==1.24.59
prometheus-client==0.14.1
python-dotenv==0.21.0
requests==2.28.1